requirements.txt
flask==2.0.1
pycryptodome==3.10.1
requests==2.26.0
beautifulsoup4==4.9.3